requirements.txt
# 项目依赖库

# Markdown解析
mistune==3.0.2

# 配置文件处理
pyyaml==6.0.1

# 进度显示
tqdm==4.66.1

# 终端颜色输出
colorama==0.4.6
